RandomSwimmingAI
RandomSwimmingAI complete.

Sets two different behaviors when it comes to rotation angles:
-If it's aggroed, allows for sharp turns.
-Else, turns are wide.

This ensures a smooth random swimming motion, avoiding sharp turns, and spinning.

Additionally, positions close to ceilings or floors are avoided as targets for RandomSwimming. If the entity ends up close to a ceiling or floor, it will force avoid it.